+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Jquery-ui replace sortablejs
+
 ## 3.9.3 - 2019-02-13
 
 ### Changed